requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
kafka-python==2.0.2
oauthlib==3.2.0
py4j==0.10.9.3
pyspark==3.2.1
requests==2.28.0
requests-oauthlib==1.3.1
tweepy==4.10.0
urllib3==1.26.9
